refatorado logica de services para auth
