+++ biblioteca/src/main/java/br/com/biblioteca/domain/authentication/AuthController.java
@@ -1,7 +1,7 @@
 package br.com.biblioteca.domain.authentication;
 
 import br.com.biblioteca.domain.user.User;
-import br.com.biblioteca.domain.user.UserService;
+import br.com.biblioteca.infrastructure.security.JwtConfig;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -11,16 +11,15 @@ import org.springframework.web.bind.annotation.*;
 @RequiredArgsConstructor
 public class AuthController {
 
-    private final UserService userService;
-    private final AuthService jwtService;
+    private final AuthService authService;
+    private final JwtConfig jwtConfig;
     private final AuthMapper authMapper;
 
-
-    @PostMapping("/login")
+    @PostMapping()
     public ResponseEntity<AuthDTO> login(@RequestBody AuthRequest authRequest) {
-        User user = userService.authenticateUser(authRequest.getEmail(), authRequest.getPassword());
+        User user = authService.authenticateUser(authRequest.getEmail(), authRequest.getPassword());
 
-        String token = jwtService.generateToken(user.getEmail(), user.getRole().name());
+        String token = jwtConfig.generateToken(user.getEmail(), user.getRole().name());
 
         AuthDTO response = authMapper.toAuthDTO(user);
         response = new AuthDTO(response.idUser(), token, response.email(), response.role(), "Login successful");
